requirements.txt
Flask==2.3.3
requests==2.31.0
rasa==3.6.0
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2
spacy==3.6.1
PyYAML==6.0
Werkzeug==2.3.7